Move i18n out of assets and support Korean
