+++ src/i18n/languages/index.ts
@@ -1,0 +1,2 @@
+export * from './en';
+export * from './ko';
