+++ src/i18n/types.ts
